requirements.txt
selenium==3.141.0
selenium-wire==4.6.3
requests2==2.16.0
pause==0.3
pytz==2022.1
opencv-python-headless==4.5.5.64
pytesseract==0.3.9
numpy == 1.22.
matplotlib == 3.5.1
ddddocr == 1.4.3